eval - check if weights file exists first
Former-commit-id: e60552ffd658189c3e337f823165c7563cff900b

+++ dsrnngan/evaluation.py
@@ -271,45 +271,47 @@ def rank_metrics_by_time(*,
 
     for model_number in model_numbers:
         gen_weights_file = os.path.join(weights_dir, "gen_weights-{:07d}.h5".format(model_number))
-        print(gen_weights_file)
+        if not os.path.isfile(gen_weights_file):
+            print(gen_weights_file, "not found, skipping")
+        else:
+            print(gen_weights_file)
+            gen.load_weights(gen_weights_file)
+            ranks, crps_scores = ensemble_ranks(mode,
+                                                gen,
+                                                batch_gen_valid,
+                                                noise_channels=noise_channels,
+                                                latent_variables=latent_variables,
+                                                batch_size=batch_size,
+                                                num_batches=num_batches,
+                                                add_noise=add_noise,
+                                                rank_samples=rank_samples,
+                                                noise_factor=noise_factor,
+                                                load_full_image=load_full_image)
+            KS = rank_KS(ranks)
+            CvM = rank_CvM(ranks)
+            DKL = rank_DKL(ranks)
+            OP = rank_OP(ranks)
+            CRPS = crps_scores.mean()
+            mean = ranks.mean()
+            std = ranks.std()
 
-        gen.load_weights(gen_weights_file)
-        ranks, crps_scores = ensemble_ranks(mode,
-                                            gen,
-                                            batch_gen_valid,
-                                            noise_channels=noise_channels,
-                                            latent_variables=latent_variables,
-                                            batch_size=batch_size,
-                                            num_batches=num_batches,
-                                            add_noise=add_noise,
-                                            rank_samples=rank_samples,
-                                            noise_factor=noise_factor,
-                                            load_full_image=load_full_image)
-        KS = rank_KS(ranks)
-        CvM = rank_CvM(ranks)
-        DKL = rank_DKL(ranks)
-        OP = rank_OP(ranks)
-        CRPS = crps_scores.mean()
-        mean = ranks.mean()
-        std = ranks.std()
+            log_line(log_fname, "{} {:.6f} {:.6f} {:.6f} {:.6f} {:.6f} {:.6f} {:.6f}".format(model_number, KS, CvM, DKL, OP, CRPS, mean, std))
 
-        log_line(log_fname, "{} {:.6f} {:.6f} {:.6f} {:.6f} {:.6f} {:.6f} {:.6f}".format(model_number, KS, CvM, DKL, OP, CRPS, mean, std))
+            # quasi-random selection to avoid generating loads of data
+            ranks_to_save = [124800, 198400, 240000, 320000]
+            # save one directory up from model weights, in same dir as logfile
+            ranks_folder = os.path.dirname(log_fname)
 
-        # quasi-random selection to avoid generating loads of data
-        ranks_to_save = [124800, 198400, 240000, 320000]
-        # save one directory up from model weights, in same dir as logfile
-        ranks_folder = os.path.dirname(log_fname)
-
-        if model_number in ranks_to_save:
-            if add_noise is False and load_full_image is False:
-                fname = 'ranks-{}.npz'.format(model_number)
-            elif add_noise is True and load_full_image is False:
-                fname = 'ranks-noise-{}.npz'.format(model_number)
-            elif add_noise is False and load_full_image is True:
-                fname = 'ranks-full_image-{}.npz'.format(model_number)
-            elif add_noise is True and load_full_image is True:
-                fname = 'ranks-full_image-noise-{}.npz'.format(model_number)
-            np.savez(os.path.join(ranks_folder, fname), ranks)
+            if model_number in ranks_to_save:
+                if add_noise is False and load_full_image is False:
+                    fname = 'ranks-{}.npz'.format(model_number)
+                elif add_noise is True and load_full_image is False:
+                    fname = 'ranks-noise-{}.npz'.format(model_number)
+                elif add_noise is False and load_full_image is True:
+                    fname = 'ranks-full_image-{}.npz'.format(model_number)
+                elif add_noise is True and load_full_image is True:
+                    fname = 'ranks-full_image-noise-{}.npz'.format(model_number)
+                np.savez(os.path.join(ranks_folder, fname), ranks)
 
 
 # def rank_metrics_by_noise(filename,
@@ -601,20 +603,22 @@ def quality_metrics_by_time(*,
 
     for model_number in model_numbers:
         gen_weights_file = os.path.join(weights_dir, "gen_weights-{:07d}.h5".format(model_number))
-        print(gen_weights_file)
+        if not os.path.isfile(gen_weights_file):
+            print(gen_weights_file, "not found, skipping")
+        else:
+            print(gen_weights_file)
+            gen.load_weights(gen_weights_file)
+            mae, rmse, ssim, lsd = image_quality(mode,
+                                                 gen,
+                                                 batch_gen_valid,
+                                                 noise_channels=noise_channels,
+                                                 latent_variables=latent_variables,
+                                                 batch_size=batch_size,
+                                                 num_instances=1,
+                                                 num_batches=num_batches,
+                                                 load_full_image=load_full_image)
 
-        gen.load_weights(gen_weights_file)
-        mae, rmse, ssim, lsd = image_quality(mode,
-                                             gen,
-                                             batch_gen_valid,
-                                             noise_channels=noise_channels,
-                                             latent_variables=latent_variables,
-                                             batch_size=batch_size,
-                                             num_instances=1,
-                                             num_batches=num_batches,
-                                             load_full_image=load_full_image)
-
-        log_line(log_fname, "{} {:.6f} {:.6f} {:.6f} {:.6f}".format(model_number, rmse.mean(), ssim.mean(), np.nanmean(lsd), mae.mean()))
+            log_line(log_fname, "{} {:.6f} {:.6f} {:.6f} {:.6f}".format(model_number, rmse.mean(), ssim.mean(), np.nanmean(lsd), mae.mean()))
 
 
 # def quality_metrics_table(weights_fname,
